Thread pool is now working.
The issue was that shutdown wasn't working properly. Initially, I was
storing the Sender half of the shutdown channel directly in the `Shared`
struct, and when passing it to a new thread, cloning the value.

This cloned value would then get dropped when the thread exited its core
run loop. But this wouldn't wake up the `recv()` call, since the
originl instace on the `Shared` struct still is in scope and hsn't been
dropped.

By wrapping this in an `Option` and setting it to `None` we trigger the
trop of the inner value, which in this case is the `Sender`, which in
turn wakes up the `recv()`. Since we cannot selectively drop parts of a
struct, wrapping it in an `Option` effectively allows us to do so.

On top of this, there's the aspect that this drop needs to happen from
the same thread as the one which is calling `recv` since the first
thread to enter the `shutdown` lock, will be responsible for waiting on
every other thread to finish.