Handle multiple pool handles
We could have multiple pool handles live across different threads that
could go out of scope at different times. We need to ensure that
`shutdown` only happens when the _last handle_ is dropped.

In order to account for this, I added a ref count which keeps track of
the handles and we use it to hold up this invariant.