Start implementing sans-io compactor.
The compaction algorithm has non-determinism in the form of the
timeouts along which it will be chceking for available files to compact.
Plus, the compaction engine can choose to buffer writes to the file
itself internally to increase throughput.

In order to ensure the core compaction protocol is deterministic to
allow thorough testability, I'm implementing this with a sans-io
approach.