Update test file system to enable introspection.
In order to observe expected behavior within the file system, we need to
be able to provide a handle and some methods to check the state of the
system after the test is conducted.

This commit instroduces a new method `new_with_fs_impl` which allows the
caller to pass in a fs object while retaining a handle to it allowing to
implement + call introspection methods in test.

This should work in theory, but the problem is that there is a deadlock
in the filesystem itself; so there's that.