Fix FS active file swap.
Tests work kids! This simple test identified a core issue with the FS
module, wherein after swapping the active file, we never reset the
state.

This commit fixes this issue by ensuring that we reset it when we have
the lock in `swap_active`.

On top of that, we also move the `active_fd` field into the `FsInner`
struct since we want to be able to reset it at the same time as we are
swapping the active file.