Store active file size in the Fs module.
This was causing a double borrow of the inner implementation, which in
the case of the test fs was causing a panic. Not entirely sure why, but
removing it fixed the issue.

Now the next problem is that the number of data files created by the
test is inconsistent across runs.